+++ monix-reactive/jvm/src/test/scala/monix/reactive/observers/OverflowStrategyFailConcurrencySuite.scala
@@ -328,7 +328,7 @@ object OverflowStrategyFailConcurrencySuite extends BaseConcurrencySuite {
       Fail(10000)
     )
 
-    (0 until 9999).foreach { x => buffer.onNext(x) }
+    (0 until 9999).foreach { x => buffer.onNext(x.toLong) }
     buffer.onComplete()
     startConsuming.success(Continue)
 
@@ -353,7 +353,7 @@ object OverflowStrategyFailConcurrencySuite extends BaseConcurrencySuite {
       Fail(10000)
     )
 
-    (0 until 9999).foreach { x => buffer.onNext(x); () }
+    (0 until 9999).foreach { x => buffer.onNext(x.toLong); () }
     buffer.onComplete()
 
     assert(complete.await(15, TimeUnit.MINUTES), "complete.await should have succeeded")
@@ -378,7 +378,7 @@ object OverflowStrategyFailConcurrencySuite extends BaseConcurrencySuite {
       Fail(10000)
     )
 
-    (0 until 9999).foreach { x => buffer.onNext(x); () }
+    (0 until 9999).foreach { x => buffer.onNext(x.toLong); () }
     buffer.onError(new RuntimeException)
     startConsuming.success(Continue)
 
@@ -403,7 +403,7 @@ object OverflowStrategyFailConcurrencySuite extends BaseConcurrencySuite {
       Fail(10000)
     )
 
-    (0 until 9999).foreach { x => buffer.onNext(x); () }
+    (0 until 9999).foreach { x => buffer.onNext(x.toLong); () }
     buffer.onError(new RuntimeException)
 
     assert(complete.await(15, TimeUnit.MINUTES), "complete.await should have succeeded")
