+++ monix-reactive/shared/src/test/scala/monix/reactive/internal/operators/DebounceRepeatedSuite.scala
@@ -30,7 +30,7 @@ object DebounceRepeatedSuite extends BaseOperatorSuite {
 
     val count = sourceCount
     val sum = sourceCount
-    Sample(o, count, sum, 1.second, 1.second)
+    Sample(o, count, sum.toLong, 1.second, 1.second)
   }
 
   def observableInError(sourceCount: Int, ex: Throwable) = None
