+++ monix-reactive/shared/src/test/scala/monix/reactive/internal/operators/DistinctUntilChangedSuite.scala
@@ -24,7 +24,7 @@ import scala.concurrent.duration.Duration.Zero
 object DistinctUntilChangedSuite extends BaseOperatorSuite {
   def createObservable(sourceCount: Int) = Some {
     val o = Observable
-      .range(0, sourceCount)
+      .range(0L, sourceCount.toLong)
       .flatMap(i => Observable.fromIterable(Seq(i, i, i)))
       .distinctUntilChanged
