+++ monix-reactive/shared/src/test/scala/monix/reactive/internal/operators/DropFirstSuite.scala
@@ -25,13 +25,13 @@ object DropFirstSuite extends BaseOperatorSuite {
   def createObservable(sourceCount: Int) = {
     require(sourceCount > 0, "sourceCount should be strictly positive")
     Some {
-      val o = Observable.range(1, sourceCount * 2).drop(sourceCount - 1)
+      val o = Observable.range(1L, sourceCount.toLong * 2).drop(sourceCount - 1)
       Sample(o, count(sourceCount), sum(sourceCount), Zero, Zero)
     }
   }
 
   def sum(sourceCount: Int): Long =
-    (1 until sourceCount * 2).drop(sourceCount - 1).sum
+    (1 until sourceCount * 2).drop(sourceCount - 1).sum.toLong
 
   def count(sourceCount: Int) =
     sourceCount
