+++ monix-reactive/shared/src/test/scala/monix/reactive/internal/operators/MergeManySuite.scala
@@ -26,7 +26,7 @@ import scala.concurrent.duration.Duration.Zero
 object MergeManySuite extends BaseOperatorSuite {
   def createObservable(sourceCount: Int) = Some {
     val o = Observable
-      .range(0, sourceCount)
+      .range(0L, sourceCount.toLong)
       .mergeMap(i => Observable.fromIterable(Seq(i, i, i, i)))
     Sample(o, count(sourceCount), sum(sourceCount), Zero, Zero)
   }
